Default viewer armature with volume settings fixed
this armature was created in an alpha version of a free plugin called OpenSim Life Tools, by Tom Ernst. I fixed some of the volume custom settings, added a non-deform PelvisInvert bone to make testing mPelvis and PELVIS weights easier, moved all the bones to one layer, and simplified the bone groups. It has tested well with the last upload of the Maxine mesh head and body. This is what I like to work with, you may want to make your own changes.